Fix feature to get train timetable for all days
Note: the bug was that the earlier method for parsing the
train timetable JSON didn't always work correctly, because
the JSON with the train timetable for a specific day don't
have the 'days' node that exists in the JSON with the train
timetable for all days. Thus, the application didn't get and
display the days of train departures if the user made a
request to receive a timetable for all days. Also such a
request returns the 'departure' and 'arrival' field with
data that the date and time formatting method wasn't parsing
correctly. These issues resulted in a bug where not all the
necessary data was displayed and timetable entries were the
same because of the missing days and the same time of departure
and arrival. That's why the bot didn't want to turn over the
pages of the request.

* fix(Timetable): remove the 'startDate' field because it's
                  redundantly

* feature(Timetable): add the 'days' field to store
                      departure days in the train timetable
                      for all days

* fix(YandexTimetableService): fix JSON train timetable parsing
                               method. Now train timetable JSON
                               for all days is parsed correctly,
                               as for one day

* fix(DateUtils): adapt the date and time firmatting method
                  for 12:00:00 format

* style(MessageUtils): 'не указана' string in method to generate
                       text of confirmation request message
                       changed to 'все дни'